Perform a more thorough search for analyzer assemblies
