+++ src/Utilities/AssemblyResolver.cs
@@ -1,0 +1,62 @@
+﻿// Copyright (c) Microsoft.  All Rights Reserved.  Licensed under the MIT license.  See License.txt in the project root for license information.
+
+using System.Collections.Generic;
+using System.IO;
+using System.Reflection;
+using System.Runtime.Loader;
+
+namespace Microsoft.CodeAnalysis.Tools.Utilities
+{
+    internal static class AssemblyResolver
+    {
+        private static readonly string[] s_extensions = new[] { "ni.dll", "ni.exe", "dll", "exe" };
+
+        internal static Assembly? TryResolveAssemblyFromPaths(AssemblyLoadContext context, AssemblyName assemblyName, string searchPath, Dictionary<string, Assembly>? knownAssemblyPaths = null)
+        {
+            foreach (var cultureSubfolder in string.IsNullOrEmpty(assemblyName.CultureName)
+                // If no culture is specified, attempt to load directly from
+                // the known dependency paths.
+                ? new[] { string.Empty }
+                // Search for satellite assemblies in culture subdirectories
+                // of the assembly search directories, but fall back to the
+                // bare search directory if that fails.
+                : new[] { assemblyName.CultureName, string.Empty })
+            {
+                foreach (var extension in s_extensions)
+                {
+                    var candidatePath = Path.Combine(
+                        searchPath, cultureSubfolder, $"{assemblyName.Name}.{extension}");
+
+                    var isAssemblyLoaded = knownAssemblyPaths?.ContainsKey(candidatePath) == true;
+                    if (isAssemblyLoaded || !File.Exists(candidatePath))
+                    {
+                        continue;
+                    }
+
+                    var candidateAssemblyName = AssemblyLoadContext.GetAssemblyName(candidatePath);
+                    if (candidateAssemblyName.Version < assemblyName.Version)
+                    {
+                        continue;
+                    }
+
+                    try
+                    {
+                        return context.LoadFromAssemblyPath(candidatePath);
+                    }
+                    catch
+                    {
+                        if (assemblyName.Name != null)
+                        {
+                            // We were unable to load the assembly from the file path. It is likely that
+                            // a different version of the assembly has already been loaded into the context.
+                            // Be forgiving and attempt to load assembly by name without specifying a version.
+                            return context.LoadFromAssemblyName(new AssemblyName(assemblyName.Name));
+                        }
+                    }
+                }
+            }
+
+            return null;
+        }
+    }
+}
